implemented UserServiceTest and RouterTest

+++ app/Services/UserService.php
@@ -1,16 +1,30 @@
 <?php
 
 namespace App\Services;
+
 use App\Models\User;
+use PDOException;
 
 class UserService extends Service {
     
     public function createUser(User $user = null) {
         $query = $this->db->prepare("INSERT INTO users VALUES (NULL, ?, ?, NOW())");
-        $query->bindParam(1, $user->data["name"], \PDO::PARAM_STR);
-        $query->bindParam(2, $user->data["email"], \PDO::PARAM_STR);
+        $query->bindParam(1, $user->name, \PDO::PARAM_STR);
+        $query->bindParam(2, $user->email, \PDO::PARAM_STR);
+        
         if(!$query->execute()) {
-            throw new \PDOException($this->getDbError($query));
+            throw new PDOException($this->getDbError($query));
+        } else {
+            $query = $this->db->prepare("SELECT id, putdate FROM users WHERE name = ? AND email = ?");
+            $query->bindParam(1, $user->name, \PDO::PARAM_STR);
+            $query->bindParam(2, $user->email, \PDO::PARAM_STR);
+            if(!$query->execute()) {
+                throw new PDOException($this->getDbError($query));
+            }
+            $result = $query->fetch(\PDO::FETCH_ASSOC);
+            
+            $user->id = $result["id"];
+            $user->putdate = $result["putdate"];
         }
         
         return $user;
@@ -20,7 +34,7 @@ class UserService extends Service {
         $query = $this->db->prepare("SELECT * FROM users WHERE id = ?");
         $query->bindParam(1, $id, \PDO::PARAM_INT);
         if(!$query->execute()) {
-            throw new \PDOException($this->getDbError($query));
+            throw new PDOException($this->getDbError($query));
         }
         
         $data = $query->fetch(\PDO::FETCH_ASSOC);
@@ -30,25 +44,30 @@ class UserService extends Service {
     
     public function updateUser(User $user = null) {
         $query = $this->db->prepare("UPDATE users SET name = ?, email = ? WHERE id = ?");
-        $query->bindParam(1, $user->data['name'], \PDO::PARAM_STR);
-        $query->bindParam(2, $user->data['email'], \PDO::PARAM_STR);
-        $query->bindParam(3, $user->data['id'], \PDO::PARAM_INT);
+        $query->bindParam(1, $user->name, \PDO::PARAM_STR);
+        $query->bindParam(2, $user->email, \PDO::PARAM_STR);
+        $query->bindParam(3, $user->id, \PDO::PARAM_INT);
         if(!$query->execute()) {
-            throw new \PDOException($this->getDbError($query));
+            throw new PDOException($this->getDbError($query));
         }
         
         return $user;
         
     }
     
-    public function deleteUser(\User $user = null) {
-        $query = $this->db->prepare("DELETE FROM users WHERE id = ?");
-        $query->bindParam(1, $user->id, \PDO::PARAM_INT);
-        if(!$query->execute()) {
-            throw new \PDOException($this->getDbError($query));
+    public function deleteUser($id) {
+        if(gettype($id) === 'integer'){
+            $query = $this->db->prepare("DELETE FROM users WHERE id = ?");
+            $query->bindParam(1, $id, \PDO::PARAM_INT);
+            $query->execute();
+            if($query->rowCount() == 0) {
+                throw new PDOException('Invalid query');
+            }
+        } else {
+            throw new PDOException('id need to be an integer');
         }
         
-        return $user;
+        return true;
         
     }
